+++ CHANGES.md
@@ -46,6 +46,7 @@ Changes in 10.0
 - SQL change program column type from text to varchar(100) NOT NULL to match with MySQL index limitation in rosariosis.sql
 - Rename YEAR_MONTH column alias to YEAR_MONTH_DATE: reserved keyword in MySQL in Dashboard.inc.php
 - SQL use DAYOFWEEK() for MySQL or extract(DOW) for PostrgeSQL, program wide
+- SQL cast custom_fields ID as VARCHAR(10) instead of TEXT for MySQL compatibility in GetStaffList.fnc.php, GetStuList.fnc.php & Search.fnc.php
 
 Changes in 9.3
 --------------